Add files via upload
AssignmentTwo is used to demonstrate and test different features related to rides. The program contains multiple sections, each of which performs specific tasks such as managing visitor queues, recording the history of rides, sorting, and exporting data. By creating the Employee and Ride objects and interacting with the Visitor object, the program shows how to apply these capabilities in a real scenario.